% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_cpdb.R
\name{combine_cpdb}
\alias{combine_cpdb}
\title{combine multiple results from cellphonedb.}
\usage{
combine_cpdb(...)
}
\arguments{
\item{...}{multiple cellphonedb outputs.}
}
\value{
combine results from multiple cellphonedb runs.
}
\description{
combine multiple results from cellphonedb.
}
