% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_cpdb.R
\name{compare_cpdb}
\alias{compare_cpdb}
\title{Plotting cellphonedb results}
\usage{
compare_cpdb(
  meta,
  celltypes,
  method = c("ttest", "wilcox", "lme"),
  p.adjust.method = "fdr",
  groupby = NULL,
  formula = NULL,
  BPPARAM = SerialParam(),
  verbose = TRUE
)
}
\arguments{
\item{meta}{data.frame containing the sample name, cellphonedb 'out' folder path (containing means.txt and pvalues.txt), and single-cell object file path (.h5ad or .rds)}

\item{celltypes}{subset celltypes for comparison}

\item{method}{one of 'ttest', 'wilcox', 'lme'}

\item{p.adjust.method}{defaults to p.adjust methods}

\item{groupby}{for significance testing. only if method is t.test or wilcoxon.}

\item{formula}{for signfiicance testing. only if method is lme.}

\item{BPPARAM}{BiocParallelParam class.}

\item{verbose}{Whether or not to print messages.}

\item{...}{passes arguments to significance tests (pairwise.t.test, pairwise.wilcox.text, lmerTest::lmer)}
}
\value{
results for plotting
}
\description{
Plotting cellphonedb results
}
\examples{
\donttest{}
}
