% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationSpot.R
\name{correlationSpot}
\alias{correlationSpot}
\title{Plotting correlations on Vissium data in Seurat}
\usage{
correlationSpot(
  st,
  genes = NULL,
  celltypes = NULL,
  geneset = NULL,
  mode = c("high", "low", "both"),
  cutoff = 0.5,
  standardize = TRUE,
  dims = 1:30,
  k.params = 10,
  resolution = 1,
  rna_slot = "SCT",
  label_slot = "predictions",
  by = c("image", "expression"),
  average_by_cluster = FALSE,
  ...
)
}
\arguments{
\item{st}{spatial transcriptomics data in Seurat.}

\item{genes}{gene or genes of interest for performing correlations. Must exist as row name(s) in the `rna_slot`.}

\item{celltypes}{celltype or celltypes of interest for performing correlations. Must exist as row name(s) in the `label_slot`.}

\item{geneset}{geneset or column in meta.data for performing correlations. Must exist as column name(s) in the meta.data.}

\item{mode}{whether or not to restrict the output to just high expressing values, low expressing values or both. Caveat with using both is low expressing ~ low expressing will still return a high correlation value.}

\item{cutoff}{percentile cut off for determining mode output.}

\item{standardize}{whether or not to scale values from 0 to 1 before performing correlations.}

\item{dims}{number of dimensions used for calculating neighborhoods in spatial data.}

\item{k.params}{number of k neighbors for calculating neighborhoods in spatial data.}

\item{resolution}{resolution of spatial clustering.}

\item{rna_slot}{name of gene expression slot. Defaults to 'SCT'.}

\item{label_slot}{name of label/prediction slot. Defaults to 'predictions'.}

\item{by}{whether or not to define spatial clusters based on image spatial location or gene expression.}

\item{average_by_cluster}{whether or not to return the output averaged across clusters.}

\item{...}{passed to Seurat::SpatialFeaturePlot}
}
\value{
SpatialFeaturePlot
}
\description{
Plotting correlations on Vissium data in Seurat
}
