% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneDotPlot.R
\name{geneDotPlot}
\alias{geneDotPlot}
\title{Plotting genes as dotplot}
\usage{
geneDotPlot(
  scdata,
  idents,
  genes,
  split.by = NULL,
  pct.threshold = 0.05,
  scaled = TRUE,
  keepLevels = TRUE,
  save.plot = FALSE,
  h = 5,
  w = 5,
  filepath = NULL,
  filename = NULL,
  heat_cols = rev(RColorBrewer::brewer.pal(9, "RdBu")),
  col_limits = NULL
)
}
\arguments{
\item{scdata}{single-cell data. can be seurat/summarizedexperiment object}

\item{idents}{column name holding the idents for each cell}

\item{genes}{genes you want to plot}

\item{split.by}{column name in the metadata/coldata table to split the spots by. If not provided, it will plot via idents provided.}

\item{pct.threshold}{float. required to keep gene expressed by minimum percentage of cells}

\item{scaled}{logical. scale the expression or not}

\item{keepLevels}{logical. keep the factor of the levels of the idents (for plotting)}

\item{save.plot}{logical. will try to save the pdf}

\item{h}{height of plot}

\item{w}{width of plot}

\item{filepath}{path to file, or path to folder}

\item{filename}{path to file}

\item{heat_cols}{colour gradient for the dot plot}

\item{col_limits}{set limits to the color gradient}
}
\value{
ggplot dot plot object of selected genes
}
\description{
Plotting genes as dotplot
}
\examples{
geneDotPlot(scdata, as.factor(scdata$split), genes = c("CD68","HLA-DRA"), save.plot = FALSE, groups = c("normal","tumor"))
}
