% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneDotPlot2.r
\name{geneDotPlot2}
\alias{geneDotPlot2}
\title{Plotting genes as dotplot}
\usage{
geneDotPlot2(scdata, idents, genes, groups, pct.threshold = 0.05,
  save.plot = TRUE, h = 5, w = 5, filepath = NULL,
  filename = NULL, heat_cols = rev(brewer.pal(9, "RdBu")),
  col_limits = NULL)
}
\arguments{
\item{scdata}{single-cell data. can be seurat/summarizedexperiment object}

\item{genes}{genes you want to plot}

\item{groups}{conditions you want to split the spots by}

\item{pct.threshold}{float. required to keep gene expressed by minimum percentage of cells}

\item{save.plot}{logical. will try to save the pdf}

\item{h}{height of plot}

\item{w}{width of plot}

\item{filepath}{path to file, or path to folder}

\item{filename}{path to file}

\item{heat_cols}{colour gradient for the dot plot}

\item{col_limits}{set limits to the color gradient}

\item{ident}{vector holding the idents for each cell}
}
\value{
ggplot dot plot object of selected genes
}
\description{
Plotting genes as dotplot
}
\examples{
geneDotPlot(scdata, as.factor(scdata$split), genes = c("CD68","HLA-DRA"), save.plot = FALSE, groups = c("normal","tumor"))
}
