% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_flat_violin.R
\name{geom_flat_violin}
\alias{geom_flat_violin}
\title{geom_flat_violin}
\source{
\url{https://gist.githubusercontent.com/benmarwick/2a1bb0133ff568cbe28d/raw/fb53bd97121f7f9ce947837ef1a4c65a73bffb3f/geom_flat_violin.R}
}
\usage{
geom_flat_violin(
  mapping = NULL,
  data = NULL,
  stat = "ydensity",
  position = "dodge",
  trim = TRUE,
  scale = "area",
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{see ggplot2::layer}

\item{data}{see ggplot2::layer}

\item{stat}{see ggplot2::layer}

\item{position}{see ggplot2::layer}

\item{trim}{see ggplot2::layer}

\item{scale}{see ggplot2::layer}

\item{show.legend}{see ggplot2::layer}

\item{inherit.aes}{see ggplot2::layer}

\item{...}{passed to ggplot2::layer}
}
\value{
geom_flat_violin
}
\description{
geom_flat_violin
}
\examples{
\donttest{
data(iris)
ggplot(iris, aes(Species, Sepal.Length)) +
    geom_flat_violin() +
    coord_flip()
}
}
