% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_flat_violin.R
\name{\%||\%}
\alias{\%||\%}
\title{geom_flat_violin}
\usage{
a \%||\% b
}
\arguments{
\item{mapping}{see ggplot2::layer}

\item{data}{see ggplot2::layer}

\item{stat}{see ggplot2::layer}

\item{position}{see ggplot2::layer}

\item{trim}{see ggplot2::layer}

\item{scale}{see ggplot2::layer}

\item{show.legend}{see ggplot2::layer}

\item{inherit.aes}{see ggplot2::layer}

\item{...}{passed to ggplot2::layer}
}
\value{
geom_flat_violin
}
\description{
geom_flat_violin
}
\examples{
sourced from https://gist.githubusercontent.com/benmarwick/2a1bb0133ff568cbe28d/raw/fb53bd97121f7f9ce947837ef1a4c65a73bffb3f/geom_flat_violin.R
somewhat hackish solution to:
https://twitter.com/EamonCaddigan/status/646759751242620928
based mostly on copy/pasting from ggplot2 geom_violin source:
https://github.com/hadley/ggplot2/blob/master/R/geom-violin.r

ggplot(diamonds, aes(cut, carat)) +
    geom_flat_violin() +
    coord_flip()
}
