% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_ktplots.R
\name{init_ktplots}
\alias{init_ktplots}
\title{documenting ktplots shortcut}
\usage{
init_ktplots()
}
\value{
quickly redocumenting ktplots
}
\description{
documenting ktplots shortcut
}
\examples{
init_ktplots() # leaves folder, install, and change back to folder
}
