% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_ktplots.R, R/ktplots.R
\name{ktplots}
\alias{ktplots}
\alias{init_ktplots}
\title{documenting ktplots shortcut}
\usage{
init_ktplots()

ktplots()
}
\value{
for internal use only. quickly redocumenting ktplots

for internal use only. runs init ktplots
}
\description{
documenting ktplots shortcut

shortcut to initiate building of ktplots
}
\examples{
\donttest{
init_ktplots() # leaves folder, install, and change back to folder
}
\donttest{
ktplots()
}
}
