% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedVlnPlot.R
\name{modify_vlnplot}
\alias{modify_vlnplot}
\title{Plotting stacked violin plot}
\usage{
modify_vlnplot(
  obj,
  feature,
  pt.size = 0,
  plot.margin = unit(c(-0.75, 0, -0.75, 0), "cm"),
  ...
)
}
\arguments{
\item{obj}{single-cell data. can be seurat/summarizedexperiment object}

\item{feature}{genes/features to plot}

\item{pt.size}{size of dots}

\item{plot.margin}{to adjust the white space between each plot.}

\item{...}{passed to Seurat::VlnPlot}
}
\value{
stacked violin plot
}
\description{
Plotting stacked violin plot
}
\examples{
features<- c("CD79A", "MS4A1", "CD8A", "CD8B", "LYZ", "LGALS3", "S100A8", "GNLY", "NKG7", "KLRB1", "FCGR3A", "FCER1A", "CST3")
StackedVlnPlot(obj = pbmc, features = features)
}
