% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{ktplots}
\alias{ktplots}
\alias{\%nin\%}
\title{miscellaneous functions}
\usage{
a \%nin\% b
}
\arguments{
\item{a}{a valute}

\item{b}{a value}
}
\value{
miscellaneous functions
}
\description{
miscellaneous functions
}
