% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cpdb3.R
\name{plot_cpdb3}
\alias{plot_cpdb3}
\title{Plotting cellphonedb results as a chord diagram}
\usage{
plot_cpdb3(
  cell_type1,
  cell_type2,
  scdata,
  idents,
  means,
  pvals,
  deconvoluted,
  p.adjust.method = NULL,
  keep_significant_only = TRUE,
  split.by = NULL,
  standard_scale = TRUE,
  separator = NULL,
  gene_symbol_mapping = NULL,
  frac = 0.1,
  remove_self = TRUE,
  desiredInteractions = NULL,
  degs_analysis = FALSE,
  directional = 1,
  alpha = 0.5,
  edge_colors = NULL,
  grid_colors = NULL,
  show_legend = TRUE,
  legend.pos.x = 20,
  legend.pos.y = 20,
  ...
)
}
\arguments{
\item{cell_type1}{cell type 1}

\item{cell_type2}{cell type 2}

\item{scdata}{single-cell data. can be seurat/summarizedexperiment object}

\item{idents}{vector holding the idents for each cell or column name of scdata's metadata. MUST match cpdb's columns}

\item{means}{object holding means.txt from cpdb output}

\item{pvals}{object holding pvals.txt from cpdb output}

\item{deconvoluted}{object holding deconvoluted.txt from cpdb output}

\item{p.adjust.method}{correction method. p.adjust.methods of one of these options: c('holm', 'hochberg', 'hommel', 'bonferroni', 'BH', 'BY', 'fdr', 'none')}

\item{keep_significant_only}{logical. Default is FALSE. Switch to TRUE if you only want to plot the significant hits from cpdb.}

\item{split.by}{column name in the metadata/coldata table to split the spots by. Can only take columns with binary options. If specified, name to split by MUST be specified in the meta file provided to cpdb prior to analysis.}

\item{standard_scale}{logical. scale the expression to range from 0 to 1. Default is TRUE}

\item{separator}{default = NULL. separator to use to split between celltypes. Unless otherwise specified, the separator will be `>@<`. Make sure the idents and split.by doesn't overlap with this.}

\item{gene_symbol_mapping}{default = NULL.column name for rowData in sce holding the actual gene symbols if row names aren't gene symbols}

\item{frac}{default = 0.1. Minimum fraction of celtypes expressing a gene in order to keep the interaction. Gene must be expressesd >= `frac` in either of the pair of celltypes in order to keep.}

\item{remove_self}{default = TRUE. Remove self-self arcs.}

\item{desiredInteractions}{default = NULL. Specific list of celltype comparisons e.g. list(c('CD4_Tcm', 'cDC1'), c('CD4_Tcm', 'cDC2')). Also accepts a dataframe where first column is celltype 1 and 2nd column is celltype 2.}

\item{degs_analysis}{if is cellphonedb degs_analysis mode.}

\item{directional}{Whether links have directions. 1 means the direction is from the first column in df to the second column, -1 is the reverse, 0 is no direction, and 2 for two directional.}

\item{alpha}{transparency for links}

\item{edge_colors}{vector of colors for links}

\item{grid_colors}{vector of colors for grids}

\item{show_legend}{whether or not to show the legend}

\item{legend.pos.x}{x position of legend}

\item{legend.pos.y}{y position of legend}

\item{...}{passes arguments plot_cpdb}

\item{scale}{logical. scale the expression to mean +/- SD. NULL defaults to TRUE.}
}
\value{
Plotting cellphonedb results as a CellChat inspired chord diagram
}
\description{
Plotting cellphonedb results as a chord diagram
}
\examples{
\donttest{
}
}
