% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGSEA_GO.R
\name{plotGSEA_GO}
\alias{plotGSEA_GO}
\title{Plotting GSEA enrichment as dotplots}
\usage{
plotGSEA_GO(gsea, top = 20, group_ref = NULL, cols = NULL,
  newlabels = NULL)
}
\arguments{
\item{gsea}{processed gsea data table}

\item{top}{how many pathways to plot}

\item{group_ref}{if more than 2 comparisons, specify the name of the group to act as reference. Otherwise, use 1 or 2 to specify the order if the reference. Defaults to using the second level.}

\item{cols}{replacement colours}

\item{newlabels}{replacement labels}
}
\value{
ggplot dot plot object of functions
}
\description{
Plotting GSEA enrichment as dotplots
}
\examples{
plotGSEA_GO(gsea_result) + ggtitle("GSEA_Hallmarks")
}
