% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGSEA_Hallmark.R
\name{plotGSEA_Hallmark}
\alias{plotGSEA_Hallmark}
\title{Plotting GSEA enrichment as dotplots}
\usage{
plotGSEA_Hallmark(gsea, cols = NULL, newlabels = NULL)
}
\arguments{
\item{gsea}{processed gsea data table}

\item{cols}{replacement colours}

\item{newlabels}{replacement labels}
}
\value{
ggplot dot plot object of functions
}
\description{
Plotting GSEA enrichment as dotplots
}
\examples{
plotGSEA_Hallmark(gsea_result) + ggtitle("GSEA_Hallmarks")
}
