% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainCloudPlot.R
\name{rainCloudPlot}
\alias{rainCloudPlot}
\title{rainCloudPlot}
\source{
\url{https://wellcomeopenresearch.org/articles/4-63}
}
\usage{
rainCloudPlot(data, groupby, parameter, default = TRUE, ...)
}
\arguments{
\item{data}{plotting data}

\item{groupby}{x axis}

\item{parameter}{y axis}

\item{default}{default plotting options}

\item{...}{passed to geom_glat_violin}
}
\value{
rainCloudPlot
}
\description{
rainCloudPlot
}
\examples{
\donttest{
data(iris)
rainCloudPlot(data = iris, groupby = "Species", parameter = "Sepal.Length") + coord_flip()
}
}
