% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedVlnPlot.R
\name{StackedVlnPlot}
\alias{StackedVlnPlot}
\title{Plotting stacked violin plot}
\source{
\url{https://divingintogeneticsandgenomics.rbind.io/post/stacked-violin-plot-for-visualizing-single-cell-data-in-seurat/}
}
\usage{
StackedVlnPlot(
  obj,
  features,
  pt.size = 0,
  plot.margin = unit(c(-0.75, 0, -0.75, 0), "cm"),
  ...
)
}
\arguments{
\item{obj}{single-cell data. can be seurat/summarizedexperiment object}

\item{features}{genes/features to plot}

\item{pt.size}{size of dots}

\item{plot.margin}{to adjust the white space between each plot.}

\item{...}{passed to Seurat::VlnPlot}
}
\value{
StackedVlnPlot
}
\description{
Plotting stacked violin plot
}
\examples{
data(kidneyimmune)
features<- c("CD79A", "MS4A1", "CD8A", "CD8B", "LYZ", "LGALS3", "S100A8", "GNLY", "NKG7", "KLRB1", "FCGR3A", "FCER1A", "CST3")
StackedVlnPlot(kidneyimmune, features = features) + ggplot2::theme(axis.text.x = ggplot2::element_text(angle = 90, hjust = 1, size = 8))
}
