% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_cpdb.R
\name{combine_cpdb}
\alias{combine_cpdb}
\title{combine multiple results from cellphonedb.}
\usage{
combine_cpdb(...)
}
\value{
combine results from multiple cellphonedb runs.
}
\description{
combine multiple results from cellphonedb.
}
\examples{
\donttest{
combine_cpdb(means, means2, means3)
combine_cpdb(pvals, pvals2, pvals3)
combine_cpdb(decon, decon2, decon3)
}
}
