% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_cpdb.R
\name{compare_cpdb}
\alias{compare_cpdb}
\title{Plotting cellphonedb results}
\usage{
compare_cpdb(
  cpdb_meta,
  sample_metadata,
  celltypes,
  celltype_col,
  groupby = NULL,
  formula = NULL,
  method = c("t.test", "wilcox", "lme"),
  BPPARAM = SerialParam(),
  version3 = FALSE,
  verbose = TRUE,
  p.adjust.mode = c("celltype", "all"),
  p.adjust.method = "fdr",
  ...
)
}
\arguments{
\item{cpdb_meta}{data.frame containing the sample name, cellphonedb 'out' folder path (containing means.txt and pvalues.txt/relevant_interactions.txt if version 3), and single-cell object file path (.h5ad or .rds)}

\item{sample_metadata}{data.frame containing the sample name, and groupings.}

\item{celltypes}{subset celltypes for comparison}

\item{celltype_col}{column name in single cell object holding celtype annotation.}

\item{groupby}{for significance testing. only if method is t.test or wilcoxon.}

\item{formula}{for signfiicance testing. only if method is lme.}

\item{method}{one of 't.test', 'wilcox', 'lme'}

\item{BPPARAM}{BiocParallelParam class.}

\item{version3}{boolean. if cellphonedb version 3}

\item{verbose}{Whether or not to print messages.}

\item{p.adjust.mode}{whether or not to correct all interactions, or to perform it on a celltype-celltype basis.}

\item{p.adjust.method}{passed to `method`in `stats::p.adjust`.}

\item{...}{passed to tests.}
}
\value{
results for plotting
}
\description{
Plotting cellphonedb results
}
\examples{
\donttest{}
}
