% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_cpdb.R
\name{compare_cpdb}
\alias{compare_cpdb}
\title{Compare cellphonedb results performed on replicates}
\usage{
compare_cpdb(
  cpdb_meta,
  sample_metadata,
  celltypes,
  celltype_col,
  groupby = NULL,
  formula = NULL,
  method = c("wilcox.test", "t.test", "lmer"),
  BPPARAM = SerialParam(),
  version3 = FALSE,
  verbose = TRUE,
  p.adjust.mode = c("celltype", "all"),
  p.adjust.method = "fdr",
  ...
)
}
\arguments{
\item{cpdb_meta}{data.frame containing the sample name, cellphonedb 'out' folder path (containing means.txt and pvalues.txt/relevant_interactions.txt if version 3), and single-cell object file path (.h5ad or .rds)}

\item{sample_metadata}{data.frame containing the sample name, and groupings.}

\item{celltypes}{subset celltypes for comparison}

\item{celltype_col}{column name in single cell object holding celtype annotation.}

\item{groupby}{for significance testing. only if method is t.test or wilcoxon.}

\item{formula}{for signfiicance testing. only if method is lme.}

\item{method}{one of 'wilcox.test', 't.test', 'lmer'}

\item{BPPARAM}{BiocParallelParam class.}

\item{version3}{boolean. if cellphonedb version 3}

\item{verbose}{Whether or not to print messages.}

\item{p.adjust.mode}{whether or not to correct all interactions, or to perform it on a celltype-celltype basis.}

\item{p.adjust.method}{passed to `method`in `stats::p.adjust`.}

\item{...}{passed to tests.}

\item{result}{output from compare_cpdb}

\item{contrast}{name of contrast to plot. only if method used was 'lmer'}

\item{groups}{name of contrast to plot. only if method used was 'lmer'}

\item{alpha}{adjusted pvalue cut off to keep in the plot.}

\item{cluster}{whether or not to perform a simple clustering in the final plot}

\item{color_limits}{upper and lower limits for plotting the results}

\item{highlight_significant}{whether or not to keep only significant results in the final plot}
}
\value{
results for plotting
}
\description{
Compare cellphonedb results performed on replicates
}
\examples{
\donttest{
data(covid_sample_metadata)
data(covid_cpdb_meta)

file <- system.file("extdata", "covid_cpdb.tar.gz", package = "ktplots")
file.copy(file, ".")
system("tar -xzf covid_cpdb.tar.gz")

covid_sample_metadata$individual <- rep(c("A", "B"), 6)
out <- compare_cpdb(cpdb_meta = covid_cpdb_meta, sample_metadata = covid_sample_metadata,
        celltypes = c("B_cell", "CD14", "CD16", "CD4", "CD8", "DCs", "MAIT", "NK_16hi",
            "NK_56hi", "Plasmablast", "Platelets", "Treg", "gdT", "pDC"), celltype_col = "initial_clustering",
        groupby = "Status_on_day_collection_summary")
plot_compare_cpdb(out)
}
}
