% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kidneyimmune}
\alias{kidneyimmune}
\alias{means}
\alias{pvals}
\alias{decon}
\alias{sig_means}
\alias{interaction_annotation}
\title{kidneyimmune}
\format{
A Seurat object with the following slots filled
\describe{
  \item{assays}{
  \itemize{Currently only contains one assay ("RNA" - scRNA-seq expression data)
  \item{data - Normalized expression data}}
}
  \item{meta.data}{Cell level metadata}
  \item{active.assay}{Current default assay}
  \item{active.ident}{Current default idents}
  \item{version}{Seurat version used to create the object}
}

Data frames containing outputs after CellPhoneDB analysis
\describe{
  \item{means}{data frame containing mean expression values for each interacting pair}
  \item{pvals}{data frame containing p values for each interacting pair}
}

An object of class \code{data.frame} with 837 rows and 11892 columns.

An object of class \code{data.frame} with 1965 rows and 115 columns.

An object of class \code{data.frame} with 837 rows and 11893 columns.

An object of class \code{data.frame} with 836 rows and 2 columns.
}
\source{
\url{https://www.kidneycellatlas.org/}
}
\usage{
data(kidneyimmune)

data(cpdb_output)

pvals

decon

sig_means

interaction_annotation
}
\description{
kidneyimmune - A small set of demo data from Stewart et al. 2019 Science. See \url{https://www.kidneycellatlas.org/}

means - Dataframe of CellPhoneDB output means.txt file

pvals - Dataframe of CellPhoneDB output pvalues.txt file

decon - Dataframe of CellPhoneDB output deconvoluted.txt file

sig_means - Dataframe of CellPhoneDB output significant_means.txt file

interaction_annotation - Example dataframe to use for interaction_grouping option in plot_cpdb2
}
\examples{
data(kidneyimmune)
data(cpdb_output)
}
\keyword{datasets}
