% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kidneyimmune}
\alias{kidneyimmune}
\alias{means}
\title{kidneyimmune}
\format{
A Seurat object with the following slots filled
\describe{
  \item{assays}{
  \itemize{Currently only contains one assay ("RNA" - scRNA-seq expression data)
  \item{data - Normalized expression data}}
}
  \item{meta.data}{Cell level metadata}
  \item{active.assay}{Current default assay}
  \item{active.ident}{Current default idents}
  \item{version}{Seurat version used to create the object}
}

Data frames containing outputs after CellPhoneDB analysis
\describe{
  \item{means}{data frame containing mean expression values for each interacting pair}
}

Data frames containing outputs after CellPhoneDB analysis
\describe{
  \item{pvals}{data frame containing p values for each interacting pair}
}
}
\source{
\url{https://www.kidneycellatlas.org/}
}
\usage{
data(kidneyimmune)

data(cpdb_output)

data(cpdb_output)
}
\description{
A small set of demo data from Stewart et al. 2019 Science

A small set of demo data from Stewart et al. 2019 Science - CellPhoneDB outut

A small set of demo data from Stewart et al. 2019 Science - CellPhoneDB outut
}
\examples{
data(kidneyimmune)
data(cpdb_output)
data(cpdb_output)
}
\keyword{datasets}
