% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ktplots.R, R/misc.R
\name{ktplots}
\alias{ktplots}
\alias{small_legend}
\alias{small_guide}
\alias{topright_legend}
\alias{topleft_legend}
\alias{bottomleft_legend}
\alias{bottomright_legend}
\alias{small_axis}
\title{shortcut to initiate building of ktplots}
\usage{
ktplots()

small_legend(fontsize = 5, ...)

small_guide(guidesize = 1, ...)

topright_legend(legendmargin = margin(6, 6, 6, 6), ...)

topleft_legend(legendmargin = margin(6, 6, 6, 6), ...)

bottomleft_legend(legendmargin = margin(6, 6, 6, 6), ...)

bottomright_legend(legendmarging = margin(6, 6, 6, 6), ...)

small_axis(fontsize = 8, linethickness = 0.1, ...)
}
\arguments{
\item{fontsize}{float/int}

\item{...}{passed to ggplot2::theme}

\item{guidesize}{float/int}

\item{legendmargin}{margin(float/int, float/int, float/int, float/int)}

\item{linethickness}{float/int}
}
\value{
runs init ktplots
}
\description{
shortcut to initiate building of ktplots
}
\examples{
ktplots()
}
