% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedVlnPlot.R, R/geom_flat_violin.R,
%   R/misc.R
\docType{data}
\name{misc}
\alias{misc}
\alias{modify_vlnplot}
\alias{extract_max}
\alias{GeomFlatViolin}
\alias{range01}
\alias{\%nin\%}
\alias{small_legend}
\alias{small_guide}
\alias{small_axis}
\alias{small_grid}
\alias{topright_legend}
\alias{topleft_legend}
\alias{bottomleft_legend}
\alias{bottomright_legend}
\title{miscellaneous functions}
\format{
An object of class \code{GeomFlatViolin} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
modify_vlnplot(
  obj,
  feature,
  pt.size = 0,
  plot.margin = unit(c(-0.75, 0, -0.75, 0), "cm"),
  ...
)

extract_max(p)

GeomFlatViolin

range01(x)

x \%nin\% y

small_legend(fontsize = 5, keysize = 0.1, marginsize = c(-0.1, 0, 0, 0), ...)

small_guide(guidesize = 1, ...)

small_axis(fontsize = 4, linethickness = 0.1, ...)

small_grid(linethickness = 0.1, panelthickness = 0.3, ...)

topright_legend(legendmargin = margin(6, 6, 6, 6), ...)

topleft_legend(legendmargin = margin(6, 6, 6, 6), ...)

bottomleft_legend(legendmargin = margin(6, 6, 6, 6), ...)

bottomright_legend(legendmargin = margin(6, 6, 6, 6), ...)
}
\arguments{
\item{obj}{single-cell data. can be seurat/summarizedexperiment object}

\item{feature}{genes/features to plot}

\item{pt.size}{size of dots}

\item{plot.margin}{to adjust the white space between each plot.}

\item{...}{passed to ggplot2::theme}

\item{p}{value}

\item{x}{a vector}

\item{y}{a vector}

\item{fontsize}{float/int}

\item{guidesize}{float/int}

\item{linethickness}{float/int}

\item{legendmargin}{margin(float/int, float/int, float/int, float/int)}
}
\value{
miscellaneous functions
}
\description{
miscellaneous functions
}
\examples{
x <-  range01(runif(100))
\donttest{
data$width <- data$width \%nin\% params$width
}
\donttest{
g + small_legend()
}
\donttest{
g + small_guide()
}
\donttest{
g + small_axis()
}
\donttest{
g + small_grid()
}
\donttest{
g + topright_legend()
}
\donttest{
g + topleft_legend()
}
\donttest{
g + bottomleft_legend()
}
\donttest{
g + bottomright_legend()
}
}
\keyword{datasets}
