% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cpdb.R
\name{plot_cpdb}
\alias{plot_cpdb}
\title{Plotting cellphonedb results}
\usage{
plot_cpdb(
  cell_type1,
  cell_type2,
  scdata,
  idents,
  means_file,
  pvals_file,
  p.adjust.method = NULL,
  keep_significant_only = FALSE,
  split.by = NULL,
  gene.family = NULL,
  genes = NULL,
  scale = NULL,
  standard_scale = NULL,
  col_option = viridis::viridis(50),
  noir = FALSE,
  highlight = "red",
  ...
)
}
\arguments{
\item{cell_type1}{cell type 1}

\item{cell_type2}{cell type 2}

\item{scdata}{single-cell data. can be seurat/summarizedexperiment object}

\item{idents}{vector holding the idents for each cell or column name of scdata's metadata. MUST match cpdb's columns}

\item{means_file}{object holding means.txt from cpdb output}

\item{pvals_file}{object holding pvals.txt from cpdb output}

\item{p.adjust.method}{correction method. p.adjust.methods of one of these options: c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none")}

\item{keep_significant_only}{logical. Default is FALSE. Switch to TRUE if you only want to plot the significant hits from cpdb.}

\item{split.by}{column name in the metadata/coldata table to split the spots by. Can only take columns with binary options. If specified, name to split by MUST be specified in the meta file provided to cpdb prior to analysis.}

\item{gene.family}{default = NULL. some predefined group of genes. can take one of these options: "chemokines", "Th1", "Th2", "Th17", "Treg", "costimulatory", "coinhibitory", "niche"}

\item{genes}{default = NULL. can specify custom list of genes if gene.family is NULL}

\item{scale}{logical. scale the expression to mean +/- SD. NULL defaults to TRUE.}

\item{standard_scale}{logical. scale the expression to range from 0 to 1. NULL defaults to FALSE.}

\item{col_option}{specify plotting colours}

\item{noir}{default = FALSE. makes it b/w}

\item{highlight}{colour for highlighting p <0.05}

\item{...}{passes arguments to grep for cell_type1 and cell_type2.}
}
\value{
ggplot dot plot object of cellphone db output
}
\description{
Plotting cellphonedb results
}
\examples{
\donttest{
data(kidneyimmune)
data(cpdb_ex)
plot_cpdb("B cell", "CD4T cell", kidneyimmune, 'celltype', means, pvals, split.by = "Experiment", genes = c("CXCL13", "CD274", "CXCR5"))
plot_cpdb("B cell", "CD4T cell", kidneyimmune, 'celltype', means, pvals, split.by = "Experiment", gene.family = 'chemokines')
}
}
