% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cpdb.R
\name{plot_cpdb}
\alias{plot_cpdb}
\title{Plotting cellphonedb results}
\usage{
plot_cpdb(cell_type1, cell_type2, means_file, pvals_file, groups,
  gene.family = NULL, genes = NULL, ...)
}
\arguments{
\item{cell_type1}{cell type 1}

\item{cell_type2}{cell type 2}

\item{means_file}{object holding means.txt from cpdb output}

\item{pvals_file}{object holding pvals.txt from cpdb output}

\item{groups}{conditions to do the plotting}

\item{gene.family}{default = NULL. some predefined group of genes. can take one of these options: "chemokines", "Th1", "Th2", "Th17", "Treg", "costimulatory", "coinhibitory", "niche"}

\item{genes}{default = NULL. can specify custom list of genes if gene.family is NULL}
}
\value{
ggplot dot plot object of cellphone db output
}
\description{
Plotting cellphonedb results
}
\examples{
pvals <- read.delim("./cpdb/output/pvalues.txt", check.names = FALSE)
means <- read.delim("./cpdb/output/means.txt", check.names = FALSE) 
plot_cpdb("Bcell", "Tcell", means, pvals, groups = c("normal", "tumor"), genes = c("CXCL13", "CD274", "CXCR5"))
}
