% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cpdb2.R
\name{plot_cpdb2}
\alias{plot_cpdb2}
\title{Plotting cellphonedb results}
\usage{
plot_cpdb2(
  cell_type1,
  cell_type2,
  scdata,
  idents,
  means,
  pvals,
  deconvoluted,
  p.adjust.method = NULL,
  keep_significant_only = FALSE,
  split.by = NULL,
  gene.family = NULL,
  genes = NULL,
  scale = NULL,
  standard_scale = NULL,
  separator = NULL,
  gene_symbol_mapping = NULL,
  frac = 0.2,
  remove_self = TRUE,
  desiredInteractions = NULL,
  interaction_grouping = NULL,
  edge_group_colors = NULL,
  node_group_colors = NULL,
  ...
)
}
\arguments{
\item{cell_type1}{cell type 1}

\item{cell_type2}{cell type 2}

\item{scdata}{single-cell data. can be seurat/summarizedexperiment object}

\item{idents}{vector holding the idents for each cell or column name of scdata's metadata. MUST match cpdb's columns}

\item{means}{object holding means.txt from cpdb output}

\item{pvals}{object holding pvals.txt from cpdb output}

\item{deconvoluted}{object holding deconvoluted.txt from cpdb output}

\item{p.adjust.method}{correction method. p.adjust.methods of one of these options: c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none")}

\item{keep_significant_only}{logical. Default is FALSE. Switch to TRUE if you only want to plot the significant hits from cpdb.}

\item{split.by}{column name in the metadata/coldata table to split the spots by. Can only take columns with binary options. If specified, name to split by MUST be specified in the meta file provided to cpdb prior to analysis.}

\item{gene.family}{default = NULL. some predefined group of genes. can take one of these options: "chemokines", "Th1", "Th2", "Th17", "Treg", "costimulatory", "coinhibitory", "niche"}

\item{genes}{default = NULL. can specify custom list of genes if gene.family is NULL}

\item{scale}{logical. scale the expression to mean +/- SD. NULL defaults to TRUE.}

\item{standard_scale}{logical. scale the expression to range from 0 to 1. NULL defaults to FALSE.}

\item{separator}{default = NULL. separator to use to split between celltypes. Unless otherwise specified, the separator will be `>@<`. Make sure the idents and split.by doesn't overlap with this.}

\item{gene_symbol_mapping}{default = NULL.column name for rowData in sce holding the actual gene symbols if row names aren't gene symbols}

\item{frac}{default = 0.2. Does nothing at the moment.}

\item{remove_self}{default = TRUE. Remove self-self arcs. might not be working.}

\item{desiredInteractions}{default = NULL. Specific list of celltype comparisons e.g. list(c('CD4_Tcm', 'cDC1'), c('CD4_Tcm', 'cDC2')). Also accepts a dataframe where first column is celltype 1 and 2nd column is celltype 2}

\item{interaction_grouping}{default = NULL. dataframe specifying groupings of cellphonedb interactions. First column must be cellphonedb's interacting_pair column. second column is whatever grouping you want.}

\item{edge_group_colors}{default = NULL. vector for colour mapping for edge groups. only used if split.by is specified.}

\item{node_group_colors}{default = NULL. vector for colour mapping for node labels.}

\item{...}{passes arguments to grep for cell_type1 and cell_type2.}
}
\value{
ggplot dot plot object of cellphone db output
}
\description{
Plotting cellphonedb results
}
\examples{
\donttest{
}
}
