% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainCloudPlot.R
\name{rainCloudPlot}
\alias{rainCloudPlot}
\title{rainCloudPlot}
\usage{
rainCloudPlot(data, groupby, parameter, default = TRUE, ...)
}
\arguments{
\item{data}{plotting data}

\item{parameter}{y axis}

\item{default}{default plotting options}

\item{...}{passed to geom_glat_violin}

\item{grouby}{x axis}
}
\value{
rainCloudPlot
}
\description{
rainCloudPlot
}
\examples{
rainCloudPlot(diamonds, cut, carat) + coord_flip()
}
